+++ R/confusion.R
@@ -29,10 +29,6 @@
 #' (\code{"adj"}, the default) or for (conditional) orientations (\code{dir}).
 #'
 #' @examples
-#' \dontrun{
-#' #############################################################################
-#' # Compare two adjacency matrices ############################################
-#' #############################################################################
 #' x1 <- matrix(c(
 #'   0, 0, 0, 0,
 #'   1, 0, 1, 0,
@@ -45,13 +41,16 @@
 #'   0, 0, 0, 0,
 #'   1, 0, 1, 0
 #' ), 4, 4, byrow = TRUE)
+#' nodes <- c("A", "B", "C", "D")
+#' rownames(x1) <- colnames(x1) <- nodes
+#' rownames(x2) <- colnames(x2) <- nodes
 #'
 #' # confusion matrix for adjacencies
 #' confusion(x2, x1)
 #'
 #' # confusion matrix for conditional orientations
 #' confusion(x2, x1, type = "dir")
-#' }
+#'
 #' @export
 confusion <- function(est_amat, true_amat, type = "adj") {
   est <- .as_tamat_any(est_amat)