+++ R/plot.R
@@ -41,6 +41,7 @@ plot.tamat <- function(x, ...) {
   plot_tempo_mech(x, ...)
 }
 
+# TODO: Fix this documentation (remove it completely?)
 #' Plot partial ancestral graph (PAG)
 #'
 #' @author This code is a modification of the fciAlgo plotting method implemented
@@ -51,29 +52,6 @@ plot.tamat <- function(x, ...) {
 #' @param ... Currently not in use.
 #'
 #' @return No return value, the function is called for its side-effects (plotting).
-#'
-#' @examples
-#' \dontrun{
-#' # simulate linear Gaussian data w unobserved variable L1
-#' n <- 100
-#' L1 <- rnorm(n)
-#' X1 <- rnorm(n)
-#' X2 <- L1 + X1 + rnorm(n)
-#' X3 <- X1 + rnorm(n)
-#' X4 <- X3 + L1 + rnorm(n)
-#' d <- data.frame(
-#'   p1_X1 = X1,
-#'   p1_X2 = X2,
-#'   p2_X3 = X3,
-#'   p2_X4 = X4
-#' )
-#'
-#' # use FCI algorithm to recover PAG
-#' res <- fci(d, test = cor_test)
-#'
-#' # plot
-#' plot(res)
-#' }
 #'
 #' @export
 plot.pag <- function(x, ...) {
@@ -112,7 +90,7 @@ plot.pag <- function(x, ...) {
   Rgraphviz::renderGraph(Rgraphviz::layoutGraph(thisg))
 }
 
-
+# TODO: Fix this documentation (remove it completely?)
 #' Plot temporal partial ancestral graph (TPAG)
 #'
 #' @author This code is a modification of the fciAlgo plotting method implemented
@@ -124,28 +102,6 @@ plot.pag <- function(x, ...) {
 #'
 #' @return No return value, the function is called for its side-effects (plotting).
 #'
-#' @examples
-#' \dontrun{
-#' # simulate linear Gaussian data w unobserved variable L1
-#' n <- 100
-#' L1 <- rnorm(n)
-#' X1 <- rnorm(n)
-#' X2 <- L1 + X1 + rnorm(n)
-#' X3 <- X1 + rnorm(n)
-#' X4 <- X3 + L1 + rnorm(n)
-#' d <- data.frame(
-#'   p1_X1 = X1,
-#'   p1_X2 = X2,
-#'   p2_X3 = X3,
-#'   p2_X4 = X4
-#' )
-#'
-#' # use FCI algorithm to recover PAG
-#' res <- tfci(d, order = c("p1", "p2"), test = cor_test)
-#'
-#' # plot
-#' plot(res)
-#' }
 #' @export
 plot.tpag <- function(x, ...) {
   plot.pag(x, ...)