+++ inst/roxygen-examples/TetradSearch_example.R
@@ -4,27 +4,24 @@
 # use the disco() or any method function, for example pc(), instead.
 
 # Requires Tetrad to be installed
-\dontrun{
-# Load data
-set.seed(1405)
-data("tpc_example")
-small_dataset <- tpc_example[1:100, ]
+if (check_tetrad_install()$installed || check_tetrad_install()$java_ok) {
+  data("tpc_example")
 
-# Recommended:
-my_pc <- pc(engine = "tetrad", test = "conditional_gaussian")
-my_pc(small_dataset)
+  # Recommended:
+  my_pc <- pc(engine = "tetrad", test = "conditional_gaussian")
+  my_pc(tpc_example)
 
-# or
-disco(data = small_dataset, method = my_pc)
+  # or
+  disco(data = tpc_example, method = my_pc)
 
-# Using R6 class:
-s <- TetradSearch$new()
+  # Using R6 class:
+  s <- TetradSearch$new()
 
-s$set_data(small_dataset)
-s$set_test(method = "conditional_gaussian", alpha = 0.05)
-s$set_alg("pc")
+  s$set_data(tpc_example)
+  s$set_test(method = "conditional_gaussian", alpha = 0.05)
+  s$set_alg("pc")
 
-g <- s$run_search()
+  g <- s$run_search()
 
-print(g)
+  print(g)
 }