+++ inst/roxygen-examples/current_heap_gb_example.R
@@ -1,6 +1,6 @@
 ### current_heap_gb() example ###
 # Returns an integer number of gigabytes on the heap (rough calculation)
-# Requires Tetrad
-\dontrun{
-causalDisco:::current_heap_gb()
+# Requires rjava to be initialized
+if (check_tetrad_install()$installed || check_tetrad_install()$java_ok) {
+  causalDisco:::current_heap_gb()
 }