+++ inst/roxygen-examples/unfreeze_example.R
@@ -3,7 +3,7 @@
 # unfreeze allows adding variables beyond the original data frame columns
 data(tpc_example)
 
-kn <- knowledge(head(tpc_example))
+kn <- knowledge(tpc_example)
 
 # this would error while frozen
 try(add_vars(kn, "new_var"))